pyproject.toml
[project]
name = "self-replicating-sytem-modeling"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "filelock>=3.20.1",
    "openai-agents>=0.6.3",
    "pydantic>=2.12.5",
    "pyyaml>=6.0.3",
]
